Write data without partition key in SplitFormat
This assumes that a) the partition key isn't part of the data, and b)
that the partition key is the first element of the record. For our
purposes, that works well enough. In the future we're probably better
off using the JsonFormat anyway, but while we're switching things out,
this should do